% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{bingo}
\alias{bingo}
\title{Menghitung probabilitas memenangkan 1 sampai semua line bingo 5x5 pada lemparan tertentu}
\usage{
bingo(lemparan, line, total_bola)
}
\arguments{
\item{lemparan}{Jumlah lemparan bola.}

\item{line}{Jumlah line bingo (minimal 1, maksimal 12).}

\item{total_bola}{Total jumlah bola (1 hingga 99).}
}
\value{
Probabilitas memenangkan seluruh line bingo.
}
\description{
Fungsi ini menghitung probabilitas memenangkan seluruh line bingo 5x5 pada lemparan tertentu.
}
\examples{
bingo_prob <- bingo(lemparan = 75, line = 12, total_bola = 99)
print(bingo_prob)
}
