% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{gjb}
\alias{gjb}
\title{Fungsi untuk menghitung gerak jatuh bebas (GJB)}
\usage{
gjb(massa, ketinggian)
}
\arguments{
\item{massa}{Massa benda (kg).}

\item{ketinggian}{Ketinggian benda (m).}
}
\value{
Kecepatan akhir GJB (m/s).
}
\description{
Fungsi ini menghitung kecepatan akhir gerak jatuh bebas (GJB) berdasarkan massa dan ketinggian benda.
}
\examples{
kecepatan_akhir_gjb <- gjb(massa = 0.2, ketinggian = 5)
print(kecepatan_akhir_gjb)

}
