% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{glbb}
\alias{glbb}
\title{Fungsi untuk menghitung gerak lurus bebas beraturan (GLBB)}
\usage{
glbb(massa, kecepatan_awal)
}
\arguments{
\item{massa}{Massa benda (kg).}

\item{kecepatan_awal}{Kecepatan awal benda (m/s).}
}
\value{
Jarak tempuh GLBB (m).
}
\description{
Fungsi ini menghitung jarak tempuh gerak lurus bebas beraturan (GLBB) berdasarkan massa dan kecepatan awal benda.
}
\examples{
jarak_tempuh_glbb <- glbb(massa = 0.5, kecepatan_awal = 10)
print(jarak_tempuh_glbb)

}
