% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{phemisphere}
\alias{phemisphere}
\title{Perimeter of Hemisphere}
\usage{
phemisphere(r)
}
\arguments{
\item{r}{is the radius of the base}
}
\value{
The perimeter of Hemisphere
}
\description{
Compute the perimeter of Hemisphere shape
}
\examples{
perimeter1 <- phemisphere(4);
perimeter2 <- phemisphere(5);
}
