\name{uaskomstat-package}
\alias{uaskomstat-package}
\alias{uaskomstat}
\docType{package}
\title{
\packageTitle{uaskomstat}
}
\description{
\packageDescription{uaskomstat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{uaskomstat}
\packageIndices{uaskomstat}

}
\author{
\packageAuthor{uaskomstat}

Maintainer: \packageMaintainer{uaskomstat}
}
\references{
Pengantar Proses Stokastik
}
\keyword{ package }
\examples{
\dontrun(data(uaskomstatdata))
phemisphere(4)
vhemisphere(5)
vspherical(5,3)
}
