\name{uaskomstat}
\alias{uaskomstat}
\docType{data}
\title{
Data untuk package uaskomstat
}
\description{
Terdapat 3 buah data yaitu data 1, data 2 dan, data 3
}
\usage{data("uaskomstat")}
\format{
  The format is:
 num [1:7, 1:3] 5 6 6 7 8 9 10 4 6 2 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:3] "data1" "data2" "data3"
}
\details{
Data 1, 2, dan 3 adalah data Random
}
\source{
dwitra.blogspot
}
\references{
Pengantar Proses Stokastik
}
\examples{
data(uaskomstat)
## maybe str(uaskomstat) ; plot(uaskomstat) ...
}
\keyword{datasets}
