% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{vhemisphere}
\alias{vhemisphere}
\title{Volume of Hemisphere}
\usage{
vhemisphere(r)
}
\arguments{
\item{r}{r is the radius of the base}
}
\value{
The volume of pyramid
}
\description{
Compute the volume of Hemisphere shape
}
\examples{
volume1 <- vhemisphere(5);
volume2 <- vhemisphere(3);
}
