% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{vspherical}
\alias{vspherical}
\title{Volume of Spherical Cap}
\usage{
vspherical(a, h)
}
\arguments{
\item{a}{a is the radius of the big ball}

\item{h}{is the height of the cap}
}
\value{
The volume of Spherical Cap
}
\description{
Compute the volume of Spherical Cap
}
\examples{
volume1 <- vspherical(5,3);
volume2 <- vspherical(4,3);
}
