\name{UASKomStat-package}
\alias{UASKomStat-package}
\alias{UASKomStat}
\docType{package}
\title{
\packageTitle{UASKomStat}
}
\description{
\packageDescription{UASKomStat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UASKomStat}
\packageIndices{UASKomStat}

}
\author{
\packageAuthor{UASKomStat}

Maintainer: \packageMaintainer{UASKomStat}
}
\references{
Komputasi Statistika
}
\keyword{ package }
\examples{
\dontrun(data(UASKomStatdata))
bingo(lemparan = 75, line = 12, total_bola = 99)
glbb(massa = 0.5, kecepatan_awal = 10)
gjb(massa = 0.2, ketinggian = 5)
}
